base/asgi.py
"""
ASGI config for base project.

It exposes the ASGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/5.0/howto/deployment/asgi/
"""

import os

from django.core.asgi import get_asgi_application
from channels.routing import ProtocolTypeRouter, URLRouter
from rooms import routings

os.environ.setdefault("DJANGO_SETTINGS_MODULE", "base.settings")

asgi_app = get_asgi_application()

application = ProtocolTypeRouter(
    {
        "http": asgi_app,
        "websocket": URLRouter([*routings.websocket_urlpatterns]),
    }
)
